+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- JSON Schema generation for torch `Device` fields
+
 ## [0.2.2] - 2026-07-02
 
 ### Changed